advanced_rag/requirements.txt
numpy>=1.26.0
chromadb>=0.5.3
pandas>=2.0.0
pypdf>=5.1.0
openai>=1.60.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.29
langchain-text-splitters>=0.0.1
umap-learn>=0.5.0
matplotlib>=3.9.0
sentence-transformers
transformers>=4.0.0
torch>=2.0.0